+++ template/src/visualization/crab_vis.rs
@@ -27,8 +27,8 @@ impl AgentRender for CrabVis {
         let agent = agent.downcast_ref::<Crab>().unwrap();
         let loc = state.field.get_location(*agent);
         match loc {
-            Some(loc) => (loc.x as f32, loc.y as f32, 0.),
-            None => (agent.loc.x as f32, agent.loc.y as f32, 0.),
+            Some(loc) => (loc.x, loc.y, 0.),
+            None => (agent.loc.x, agent.loc.y, 0.),
         }
     }
 
@@ -45,7 +45,7 @@ impl AgentRender for CrabVis {
         } else {
             concrete_agent.last_d.y.atan2(concrete_agent.last_d.x)
         };
-        (rotation - PI) as f32
+        rotation - PI
     }
 
     /// Specify the code to execute for each frame, for each agent.
